requirements.txt
google.generativeai
python-dotenv
langchain
langchain-google-genai
pypdf
streamlit